SimpleLottery: Moving init in params implementation block
