+++ src/simple_lottery/mod.rs
@@ -1,8 +1,6 @@
 //! Simple lottery construction for the Telescope ALBA scheme.
 //! Covering the Section 4.1 of the paper.
 
-pub mod init;
-
 pub mod params;
 
 pub mod proof;
